mv locale-dropdown && split it from the topbar
