+++ src/app/[locale]/layout.tsx
@@ -8,9 +8,10 @@ import { notFound } from "next/navigation";
 import Script from "next/script";
 import { Suspense } from "react";
 
-import LocaleDropdown from "./locale-dropdown";
+import LocaleDropdown from "../../components/custom/locale-dropdown";
 import Provider from "./provider";
 import { QueryProvider } from "./queryProvider";
+import TopBar from "./topbar";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -57,7 +58,7 @@ export default async function RootLayout({
           <QueryProvider>
             <Provider>
               <NextIntlClientProvider>
-                <LocaleDropdown />
+                <TopBar />
                 {children}
                 <Toaster />
               </NextIntlClientProvider>
